+++ lib/src/main/java/com/diffplug/spotless/npm/PrettierFormatterStep.java
@@ -41,8 +41,10 @@ public class PrettierFormatterStep {
 
 	public static final String NAME = "prettier-format";
 
+	public static final String DEFAULT_VERSION = "2.8.8";
+
 	public static final Map<String, String> defaultDevDependencies() {
-		return defaultDevDependenciesWithPrettier("2.8.1");
+		return defaultDevDependenciesWithPrettier(DEFAULT_VERSION);
 	}
 
 	public static final Map<String, String> defaultDevDependenciesWithPrettier(String version) {